@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 
 class SizeConfig {
-  MediaQueryData _mediaQueryData;
-  static double screenWidth;
-  static double screenHeight;
-  static double _safeAreaHorizontal;
-  static double _safeAreaVertical;
-  static double b;
-  static double v;
+  late MediaQueryData _mediaQueryData;
+  static late double screenWidth;
+  static late double screenHeight;
+  static late double _safeAreaHorizontal;
+  static late double _safeAreaVertical;
+  static late double b;
+  static late double v;
 
   void init(BuildContext context) {
     _mediaQueryData = MediaQuery.of(context);
